리뷰 댓글 사용자 정보 추가

+++ perfume-api/src/main/java/io/perfume/api/review/adapter/in/http/dto/GetReviewCommentsRequestDto.java
@@ -2,13 +2,13 @@ package io.perfume.api.review.adapter.in.http.dto;
 
 import io.perfume.api.review.application.facade.dto.ReviewCommentDetailCommand;
 
-public record GetReviewCommentsRequestDto(Long size, String before, String after) {
+public record GetReviewCommentsRequestDto(Integer size, String before, String after) {
 
   public ReviewCommentDetailCommand toCommand(long reviewId) {
     return new ReviewCommentDetailCommand(reviewId, getSizeOrDefault(), after, before);
   }
 
-  private Long getSizeOrDefault() {
+  private Integer getSizeOrDefault() {
     return size == null ? 10 : size;
   }
 }
